requirements.txt
lask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
gevent==1.0
greenlet==0.4.2
itsdangerous==0.23
ujson==1.33
